put types in a file
